requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
environs==9.3.0
geojson==2.5.0
marshmallow==3.10.0
Pillow==7.1.2
python-dotenv==0.15.0
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.7
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
environs==9.3.0
geojson==2.5.0
idna==2.10
marshmallow==3.10.0
Pillow==7.1.2
python-dotenv==0.15.0
pytz==2020.1
requests==2.25.1
sqlparse==0.3.1
urllib3==1.26.3
